feat(model): update the todos cloud functions and modularize the save/edit screens
In order to stay consistent with Firebase, the todo have to contain a Timestamp.
This caused the cloud function to fail. To overcome this issue, in the client side
of the app we handle only JS Date and when communicating with the backend, we
needed to transmit a Timestamp. This commit fix this by forcing the user of the
cloud functions to pass all arguments, and in the CF we handle the type conversion.
Moreover, this commit adds a abstractTodoEditor, which provides as it called an abstract
todo editor, which reduces significantly boilerplate code between the save and edit screens.